+++ NEWS.md
@@ -1,3 +1,11 @@
+# consort 0.2.0
+
+* No side box if no subjects excluded.
+* Add text align option for terminal box.
+* Allow continuous terminal box.
+* Exported box label generation function.
+* Added some examples.
+
 # consort 0.1.1
 
 * Initial release.